+++ HVLib/HTTP/HVHttpRequestResponse.m
@@ -59,7 +59,7 @@
 
 -(void)connectionDidFinishLoading:(NSURLConnection *)connection
 {
-    int statusCode = [((NSHTTPURLResponse *)m_response) statusCode];
+    NSInteger statusCode = [((NSHTTPURLResponse *)m_response) statusCode];
     if (statusCode != 200)
     {       
         HVHttpException* ex = [[HVHttpException alloc] initWithStatusCode:statusCode];
