+++ HVLib/Types/Base/HVDay.m
@@ -21,11 +21,11 @@
 
 @implementation HVDay
 
--(int) min
+-(NSInteger) min
 {
     return 1;
 }
--(int) max
+-(NSInteger) max
 {
     return 31;
 }
