+++ HVLib/Types/Base/HVHour.m
@@ -20,11 +20,11 @@
 
 @implementation HVHour
 
--(int) min
+-(NSInteger) min
 {
     return 0;
 }
--(int) max
+-(NSInteger) max
 {
     return 59;
 }
